Add product scan (by input) and print receipt

+++ src/Product.java
@@ -3,9 +3,9 @@ public class Product {
 
     private String idCode;
     private String name;
-    private float price;
+    private Float price;
 
-    public Product(String code, String name, int price) {
+    public Product(String code, String name, Float price) {
         this.idCode = code;
         this.name = name;
         this.price = price;
